# Add support for the native Activity properties Status and StatusDescription